+++ app/admin/ecommerce/vendoraddresses/[id]/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { DataGrid, GridColDef } from "@mui/x-data-grid";
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import { fetcher, useFetcher } from "../../../../components/global/fetcher";
 import Loading from "../../../../../components/global/loading";
 import { useAtom } from "jotai";
@@ -10,25 +10,41 @@ import LightDataGrid from "@/app/components/global/LightDataGrid/LightDataGrid";
 import { IconButton } from "@mui/material";
 import DeleteIcon from "@mui/icons-material/Delete";
 import ModeEditIcon from "@mui/icons-material/ModeEdit";
+import Swal from "sweetalert2";
+
 export default function VendorAddress({ params }) {
   const [title, setTitle] = useAtom(pageTitle);
+  const [triggered, setTriggered] = useState(false);
 
   useEffect(() => {
     setTitle({
       title: "آدرس های فروشگاه",
       buttonTitle: "افزودن آدرس جدید",
-      link: `/admin/ecommerce/vendorAddresses/${params.id}/new`,
+      link: `/admin/ecommerce/vendoraddresses/${params.id}/new`,
     });
   }, []);
 
-  const deleteGuarantee = async (id) => {
+  const deleteAddress = async (id) => {
     try {
-      const req = await fetcher({
-        url: `/v1/api/ecommerce/vendorAddresses/${id}`,
-        method: "DELETE",
+      const result = await Swal.fire({
+        title: "مطمئن هستید؟",
+        text: "با حذف این سفارش امکان بازگشت آن وجود ندارد",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "بله حذفش کن",
+        cancelButtonText: "لغو",
       });
-      toast.success("موفق");
-      refetchAddresses();
+
+      if (result.isConfirmed) {
+        const req = await fetcher({
+          url: `/v1/api/ecommerce/vendorAddresses/${id}`,
+          method: "DELETE",
+        });
+        toast.success("موفق");
+        setTriggered(!triggered);
+      }
     } catch (error) {
       toast.error(error.message);
     }
@@ -77,13 +93,13 @@ export default function VendorAddress({ params }) {
           {}
 
           <a
-            href={`/admin/ecommerce/vendoraddresses/${params.id}/edit/${row.id}`}
+            href={`/admin/ecommerce/vendoraddresses/${params.id}/edit/${row?.original?.address?.id}/${row.original.id}`}
           >
             <IconButton aria-label="delete" color="primary">
               <ModeEditIcon />
             </IconButton>
           </a>
-          <a onClick={(e) => deleteRow(row.id)}>
+          <a onClick={(e) => deleteAddress(row.original.address.id)}>
             <IconButton aria-label="delete" color="error">
               <DeleteIcon />
             </IconButton>
@@ -98,6 +114,7 @@ export default function VendorAddress({ params }) {
       <LightDataGrid
         url={`/v1/api/ecommerce/vendorAddresses?sortOrder=DESC&offset=0&limit=10&orderBy=id&vendorId=${params.id}`}
         columns={columns}
+        triggered={triggered}
       />
     </div>
   );
